+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `src/hls/autoselect.rs:31:57` panic.
+
 ## [0.2.5] - 2023-01-09
 
 ### Changed